rating and map

+++ FoodAPI/Interfaces/IMapRoutingService.cs
@@ -5,6 +5,6 @@ namespace FoodAPI.Interfaces
     public interface IMapRoutingService
     {
         Task<TravelSummary?> GetShortestDistance
-            (float latOrigin, float lngOrigin, float latArrival, float lngArrival);
+            (double latOrigin, double lngOrigin, double latArrival, double lngArrival);
     }
 }
